+++ vite.config.ts
@@ -81,6 +81,19 @@ export default defineConfig({
         },
       },
       {
+        entry: 'electron/workers/library-worker.ts',
+        onstart() {},
+        vite: {
+          build: {
+            outDir: 'dist-electron',
+            emptyOutDir: false,
+            rollupOptions: {
+              external: electronExternal,
+            },
+          },
+        },
+      },
+      {
         entry: 'electron/preload.ts',
         onstart(args) {
           args.reload()